requirements.txt
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka-python==1.4.3
MarkupSafe==1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
oauthlib==2.1.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py4j==0.10.7
Pygments==2.2.0
PySocks==1.6.8
pyspark==2.3.2
python-dateutil==2.7.5
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
requests-oauthlib==1.0.0
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
tweepy==3.6.0
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
